cambios finales v2

+++ routes/web.php
@@ -114,12 +114,13 @@ Route::prefix('admin')->group(function () {
         Route::patch('/orders/{order}', [OrderController::class, 'update'])->name('admin.orders.update');
         
         // Eliminar orden
-        Route::delete('/orders/{order}', [OrderController::class, 'destroy'])->name('orders.destroy');
+        Route::delete('/orders/{order}', [OrderController::class, 'destroy'])->name('admin.orders.destroy');
 
         //para ver y editar los pedidos personalizados
         Route::patch('/custom-orders/{customOrder}', [CustomOrderController::class, 'updateStatus'])->name('admin.orders.updateCustomStatus');
         
-        
+        // Eliminar una orden personalizada
+
         //Route::get('/catalogo', [CategoryController::class, 'index'])->name('admin.catalogo');
         //para poder guardar los cambios de modo editar
         //Route::put('/categories/{id}', [CategoryController::class, 'update'])->name('categories.update');
